Update to some Cell code
Added implementation with the Frame structs, should be easier to comprehend from the GUI standpoint

+++ Environmental-Simulation/Environmental-Simulation/EnvSim GUI/AnimalCell.cpp
@@ -176,7 +176,7 @@ double* AnimalCell::Observe()
 
 // Public functions
 
-AnimalCell::AnimalCell(AnimalModel* m, Grid* map, double xCoord, double yCoord, int cellCount)
+AnimalCell::AnimalCell(AnimalModel* m, Grid* map, int xCoord, int yCoord, int cellCount)
 {
 	model = m;
 	grid = map;
